created bin/.path/ directory the public bin dir
the public binaries that are to be on user shell should be added to
bin/.path/ as links. The actual bin/ directory has too many commands and
most of it are not useful to the end user. So having bin/path/ for user
commands is handy, specially for external packages installed, e.g. drush
(that is the actual reason that this file is being modified right now).