change htpasswd logic to better adapt to distros
with the differences with apache 2.2 and 2.4 the old htpasswd scheme didn't
work on apache 2.4, and didn't work well with cloned sites. This commit
changes the .htaccess based lock into a template access lock, in a way that
user changes to .htaccess doesn't inadvertently remove password protection,
and also cloned sites don't get the wrong paths on files.

Also fixed several small bugs on the utility functions, that were found when
developing this change.