fixed problem of vhost archive name with tar
the template name used for archiving vhost files
(config/key_value/vhost_archive_template_str) had a column(:)
in the name, separting hour and minute. When restoring it using only the
basename of file tar attempt to use the first part of the name (before the
column) as a remote host (just like on ssh). Changed the format to: HHhMMm
to indicate hour and minute.