updated default PHP version to 5.6 for CentOS 6.x
The default version of PHP for the system-wide install for the Remi
repository was updated to 5.6, what conflicted with our base install (broke
the installation of some PHP modules). Updated the default PHP repository for
Remi to 5.6 to fix it.

For CentOS 6.x the default PHP version is 5.3 that is very outdated.

So we use the Remi repository to get a more up-to-date PHP version that
is more current for the needs of current applications (as of early
2016). If the repository for the newest version is not explicitly
enabled, CentOS will get the old PHP version.