workaround to .tar files sent as public files
sometimes the import feature is breaking because a .tar file is sent (the
original .tar.gz is being modified somewhere in the system before reaching
the VPS).

With this change we use the feature of newer tar versions that automatically
detect the compression type and extract the file without explicitly
specifying the file type.