CLI: internal change in command-args processing
Process commands from the longest sequence of non-option to the shortest.

Previously the processing started with the shortest matching paths, but this
caused a conflict with similarly named commands, for example, 'enable vhost'
and 'enable vhost snapshots'. The latter forms was always unreachable
because of the shortest match was caught first.